Changes: - Finalized data structure to represent the ERD. - Made an initial insert into the database upon startup. - Added composite ID to DMNVersions. - Started with repositories

+++ src/main/java/nl/syntouch/models/composites/DMNVersionId.java
@@ -6,11 +6,11 @@ import java.util.Objects;
 
 @Embeddable
 public class DMNVersionId implements Serializable {
-    private Long dmn;
-    private Long version;
+    private Integer dmn;
+    private Integer version;
 
     public DMNVersionId() {}
-    public DMNVersionId(Long dmn, Long version) {
+    public DMNVersionId(Integer dmn, Integer version) {
         this.dmn = dmn;
         this.version = version;
     }
