+++ src/main/java/nl/syntouch/rest/DMNVersionResource.java
@@ -3,5 +3,5 @@ package nl.syntouch.rest;
 import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
 import nl.syntouch.models.DMNVersion;
 
-public interface DMNVersionResource extends PanacheEntityResource<DMNVersion, Long> {
+public interface DMNVersionResource extends PanacheEntityResource<DMNVersion, Integer> {
 }
