+++ src/main/java/nl/syntouch/rest/DeploymentResource.java
@@ -3,6 +3,6 @@ package nl.syntouch.rest;
 import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
 import nl.syntouch.models.Deployment;
 
-public interface DeploymentResource extends PanacheEntityResource<Deployment, Long> {
+public interface DeploymentResource extends PanacheEntityResource<Deployment, Integer> {
 
 }
