+++ src/main/java/nl/syntouch/rest/DomainResource.java
@@ -3,6 +3,6 @@ package nl.syntouch.rest;
 import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
 import nl.syntouch.models.Domain;
 
-public interface DomainResource extends PanacheEntityResource<Domain, Long> {
+public interface DomainResource extends PanacheEntityResource<Domain, Integer> {
 
 }
