+++ src/main/java/nl/syntouch/rest/EnvironmentResource.java
@@ -3,5 +3,5 @@ package nl.syntouch.rest;
 import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
 import nl.syntouch.models.Environment;
 
-public interface EnvironmentResource extends PanacheEntityResource<Environment, Long> {
+public interface EnvironmentResource extends PanacheEntityResource<Environment, Integer> {
 }
